# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Places",
        "description": "Documentação da API Places",
        "contact": {
            "email": "seu-email@exemplo.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registra um novo usuário",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "João da Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Autentica um usuário e retorna um token",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Autenticado com sucesso, retorna token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJK..."
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Revoga o token do usuário autenticado",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Logout feito com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/places": {
            "get": {
                "tags": [
                    "Places"
                ],
                "summary": "Lista todos os lugares",
                "operationId": "98b7b5e708ea1544e70cf25ec8621387",
                "parameters": [
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Filtra por cidade",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtra por tipo",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtra por nome",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de lugares paginada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Place"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Places"
                ],
                "summary": "Cria um novo lugar",
                "operationId": "9b1406d5892a139db93bd35d7a248450",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PlaceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lugar criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Place"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/places/{id}": {
            "get": {
                "tags": [
                    "Places"
                ],
                "summary": "Mostra detalhes de um lugar específico",
                "operationId": "4ee90b2a793be42613fcf8ae54ff0cd9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do lugar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do lugar",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Place"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lugar não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Places"
                ],
                "summary": "Atualiza um lugar",
                "operationId": "3eba858ed63a31cf3380c85c3bcd4ea2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do lugar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PlaceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lugar atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Place"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lugar não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Places"
                ],
                "summary": "Remove um lugar",
                "operationId": "0b4281911812317e70e0ed38b948fa6d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do lugar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lugar removido com sucesso"
                    },
                    "404": {
                        "description": "Lugar não encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Place": {
                "title": "Place",
                "required": [
                    "id",
                    "name",
                    "city",
                    "type"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Praia do Sol"
                    },
                    "city": {
                        "type": "string",
                        "example": "Rio de Janeiro"
                    },
                    "type": {
                        "type": "string",
                        "example": "Praia"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "PlaceRequest": {
                "title": "PlaceRequest",
                "required": [
                    "name",
                    "city",
                    "type"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Praia do Sol"
                    },
                    "city": {
                        "type": "string",
                        "example": "Rio de Janeiro"
                    },
                    "type": {
                        "type": "string",
                        "example": "Praia"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Places",
            "description": "Places"
        }
    ]
}